requirements.txt
hydra-core
numpy
matplotlib
scikit-learn
tqdm
pillow
joblib
pytorch-lightning
wandb